Architecture rework (#56)
* chore: Add Zenject dependency injection framework

* chore: Add Zenject dependency injection framework

* fix: Laser projectile refraction not working

The invalid Projectile layer was still being used

* refactor: Unify character scripts

* fix: Missing data in scene

* feat: Sprite fragmenter fragments use right sprite texture

* refactor: Abstraction over input system

Cleaned up input system, but the player side of things is still not satisfactory

* feat: Increase base player speed

* feat: Despawn sprite fragments after some time, improve ObjectInstantiator

* refactor: Service locator to inject dependencies into objects with Unity-managed lifetimes

I couldnt find a better solution because zenject cannot inject in serializereference or behavior tree actions. Compared to using static classes it's still better for testability.

* feat: Pool sprite fragments and explosion FXs

doesnt work the first time for some reason

* refactor: Make GlobalCoroutine a service

* fix: Error with globalcoroutine binding

Should be addonroot

* feat: Add unit testing setup

* fix: Sprite fragmentation not working  the first time

The pool's OnSpawned is called after the item's OnSpawned, so used an event to wait for the fragment to be fully spawned before starting destroy coroutine.

I dont really like the setup for this but it works okay lol

* feat: Add missing scene file

* fix: Taking damage from explosion not working

* refactor: IDamageable interface for stuff that can lose health

Will be needed by base +